--- views/html/newStory.html
@@ -1,105 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta http-equiv="Content-Type" content="text/html; charset=UTF-8"/>
-  <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1.0"/>
-  <title>Starter Template - Materialize</title>
-
-  <!-- CSS  -->
-  <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
-  <link href="public/css/materialize.css" type="text/css" rel="stylesheet" media="screen,projection"/>
-  <link href="public/css/style.css" type="text/css" rel="stylesheet" media="screen,projection"/>
-</head>
-<body class="grey darken-3">
-  <nav class="blue-grey" role="navigation">
-    <div class="nav-wrapper container"><a id="logo-container" href="#" class="brand-logo">[word]:[wad]</a>
-      <ul class="right hide-on-med-and-down">
-        <li><a href="#">Create Wad</a></li>
-        <li><a href="#">Login</a></li>
-        <li><a href="#">Browse</a></li>
-      </ul>
-
-      <ul id="nav-mobile" class="sidenav">
-        <li><a href="#">Navbar Link</a></li>
-      </ul>
-      <a href="#" data-target="nav-mobile" class="sidenav-trigger"><i class="material-icons">menu</i></a>
-    </div>
-  </nav>
-  <div class="section no-pad-bot grey darken-3" id="index-banner">
-    <div class="container">
-      <div class="row center">
-        <h4 class="white-text">[new]:[wad]</h4>
-      </div>
-      <br><br>
-      <div class="row">
-        <form class="col s12">
-          <div class="row">
-            <div class="input-field col s12">
-              <input id="title" type="text" class="validate white-text">
-              <label for="title">Title</label>
-            </div>
-          </div>
-          <div class="row">
-            <div class="input-field col s12">
-              <input id="genre" type="text" class="validate white-text">
-              <label for="genre">Genre</label>
-            </div>
-          </div>
-          <div class="row">
-            <form class="col s12">
-              <div class="row">
-                <div class="input-field col s12">
-                  <textarea  id="newStory" class="materialize-textarea white-text"></textarea>
-                  <label for="newStory">Begin your story</label>
-                  <span class="white-text" class="helper-text" data-error="wrong" data-success="right">the text area will expand as you type, returns are recognized</span>
-                </div>
-              </div>
-            </form>
-          </div>
-          
-          
-        </form>
-        <button id="newStorySubmit" class="btn waves-effect waves-light" type="submit" name="action">Publish
-          <i class="material-icons right">control_point</i>
-        </button>
-      </div>
-      <br><br>
-
-    </div>
-  </div>
-
-
-  
-
-  <footer class="page-footer grey darken-3">
-    <div class="container">
-      <div class="row">
-        <div class="col l6 s12">
-          <h5 class="white-text">[word]:[wad] © 2018</h5>
-          
-
-
-        </div>
-        <div class="col l3 s12">
-          <h5 class="white-text">links</h5>
-          <ul>
-            <li><a class="white-text" href="#!">create</a></li>
-            <li><a class="white-text" href="#!">browse</a></li>
-            <li><a class="white-text" href="#!">login</a></li>
-            
-          </ul>
-        </div>
-        
-      </div>
-    </div>
-    
-  </footer>
-
-
-  <!--  Scripts-->
-  <script src="https://code.jquery.com/jquery-2.1.1.min.js"></script>
-  <script src="public/js/materialize.js"></script>
-  <script src="public/js/init.js"></script>
-
-  </body>
-</html>
