--- views/html/storyFrag.html
@@ -1,142 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta http-equiv="Content-Type" content="text/html; charset=UTF-8"/>
-  <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1.0"/>
-  <title>Starter Template - Materialize</title>
-
-  <!-- CSS  -->
-  <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
-  <link href="public/css/materialize.css" type="text/css" rel="stylesheet" media="screen,projection"/>
-  <link href="public/css/style.css" type="text/css" rel="stylesheet" media="screen,projection"/>
-</head>
-<body class="grey darken-3">
-  <nav class="blue-grey" role="navigation">
-    <div class="nav-wrapper container"><a id="logo-container" href="#" class="brand-logo">[word]:[wad]</a>
-      <ul class="right hide-on-med-and-down">
-        <li><a href="#">Create Wad</a></li>
-        <li><a href="#">Login</a></li>
-        <li><a href="#">Browse</a></li>
-      </ul>
-
-      <ul id="nav-mobile" class="sidenav">
-        <li><a href="#">Navbar Link</a></li>
-      </ul>
-      <a href="#" data-target="nav-mobile" class="sidenav-trigger"><i class="material-icons">menu</i></a>
-    </div>
-  </nav>
-  <div class="section no-pad-bot grey darken-3" id="index-banner">
-    <div class="container">
-      <div class="row center">
-        <h5 class="white-text">[title]:[wad-22]</h5>
-        <h5 class="grey-text">[author]:[ryeng] [genre]:[humor]  [creadted]:[1.11.2018]  [contributors]:[1]</h4>
-      </div>
-      <br><br>
-      <p class="white-text">
-            In the deranged Halloween slice. Captive Agony deteriorated, fatal or chilling is, grotesque Halloween exorcism. Breathing heavily damp, cold and crying running, stalking slicing stabbing dripping (blood), is creaking screaming 666 at kettle. Sinister at creepy anxiety bite monster electrocution smashed in death. killer mental psychotic sliced. screams exorcism ooze.
-
-            Smashed. Breathing heavily hell and rabid anthrax fanatic tearing at a squeal. Dungeon in a gruesome, a disembowel unknown. Cat at decapitated guns. In the deranged Halloween slice. Captiv.
-            
-            Agony deteriorated, fatal or chilling is, grotesque Halloween exorcism. Suicide Slash demon at convulsing darkness a evil pain burn. Haunt suicide silent, gory as demonic alarming, buried in fallen angel. Undead obsession tear horrifying, a devil fiendish tense in. Children creepy deranged.
-            
-            Ashes a killer a, monster mental hospital. Edginess anxiety tension, claw at deteriorated in, creep a tear. Psycho crazy mental hospital moon. Disembowel stab, chains dungeon torment fiendish, gruesome psychopath monster captive, a mutilation possession obsession at fanatic. Creep cold graves, shadow non fear a, psychotic ashe.
-            
-            Ghost. Pain needles commodo drool rabid nightmare. Occult at alley pushed. Screams creaking tension kettle our exorcist. Blood ashes tearing in drowning spooky at fanatic with murderer. Witch ashes eyeball undead, in bury burn hell flames. Falling are graveyard graves est. Anthrax putrid, disturbing and morbid convulsing, rotten menacing fatal bleak, monstrosity dread wit.
-            
-            Demon scourge terrifying. Anthrax putrid, disturbing and morbid convulsing, rotten menacing fatal bleak, monstrosity dread with demon scourge terrifying. Cat at decapitated guns. Creep cold graves, shadow non fear a, psychotic ashes ghost. Haunt sliced at choking cut drool. Graves slice ominous is guns exorcism guts. Witch ashes eyeball undead, in bury burn hell flames. screams exorcism ooze.
-      </p>
-      <div class="row">
-                    <div class="col s12 m6 l4">
-                      <div class="card blue-grey darken-1">
-                        <div class="card-content white-text">
-                          <span class="card-title">Choice 1</span>
-                          <p>In the deranged Halloween slice. Captive Agony deteriorated, fatal or chilling is, grotesque Halloween exorcism. </p></div>
-                        <div class="card-action">
-                          <a href="#choice1">Choose This</a>
-                        </div>
-                      </div>
-                    </div>
-                  <div class="col s12 m6 l4">
-                        <div class="card blue-grey darken-1">
-                          <div class="card-content white-text">
-                            <span class="card-title">Choice 2</span>
-                            <p>In the deranged Halloween slice. Captive Agony deteriorated, fatal or chilling is, grotesque Halloween exorcism. </p>
-                          </div>
-                          <div class="card-action">
-                            <a href="#choice1">Choose This</a>
-                          </div>
-                        </div>
-                      </div>
-                    <div class="col s12 m6 l4">
-                            <div class="card blue-grey darken-1">
-                              <div class="card-content white-text">
-                                <span class="card-title">Empty</span>
-                                <form>
-                                        <div class="row">
-                                          <div class="input-field">
-                                            <textarea  id="newChoice" class="materialize-textarea white-text"></textarea>
-                                            <label for="newChoice">Enter you next choice.</label>
-                                            </div>
-                                        </div>
-                                      </form>
-                                </div>
-                              <div class="card-action">
-                                
-                                <div class="switch">
-                                        <label>
-                                          leave empty
-                                          <input type="checkbox">
-                                          <span class="lever"></span>
-                                          submit
-                                        </label>
-                                      </div>
-                                      <a href="#newSubmit">Submit</a>
-                              </div>
-                            </div>
-                          </div>
-                        </div>
-        
-           <!-- <button id="newStorySubmit" class="btn waves-effect waves-light" type="submit" name="action">Publish
-          <i class="material-icons right">control_point</i>
-        </button> -->
-      </div>
-      <br><br>
-
-    </div>
-  </div>
-
-
-  
-
-  <footer class="page-footer grey darken-3">
-    <div class="container">
-      <div class="row">
-        <div class="col l6 s12">
-          <h5 class="white-text">[word]:[wad] © 2018</h5>
-          
-
-
-        </div>
-        <div class="col l3 s12">
-          <h5 class="white-text">links</h5>
-          <ul>
-            <li><a class="white-text" href="#!">create</a></li>
-            <li><a class="white-text" href="#!">browse</a></li>
-            <li><a class="white-text" href="#!">login</a></li>
-            
-          </ul>
-        </div>
-        
-      </div>
-    </div>
-    
-  </footer>
-
-
-  <!--  Scripts-->
-  <script src="https://code.jquery.com/jquery-2.1.1.min.js"></script>
-  <script src="public/js/materialize.js"></script>
-  <script src="public/js/init.js"></script>
-
-  </body>
-</html>
