+++ views/layouts/main.handlebars
@@ -10,12 +10,15 @@
   <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
   <link href="/css/materialize.css" type="text/css" rel="stylesheet" media="screen,projection" />
   <link href="/css/style.css" type="text/css" rel="stylesheet" media="screen,projection" />
+  <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
+  <script src="/js/main.js"></script>
 </head>
 
 <body class="grey darken-3">
 {{{body}}}
 
   <!--  Scripts-->
+
   <script src="https://code.jquery.com/jquery-2.1.1.min.js"></script>
   <script src="public/js/materialize.js"></script>
   <script src="public/js/init.js"></script>
