+++ views/login.handlebars
@@ -21,7 +21,7 @@
     </div>
     <br><br>
     <div class="row">
-      <form class="col s12">
+      <form id = "userForm" class="col s12">
         <div class="row">
           <div class="input-field col s12">
             <input id="userName" type="text" class="validate">
@@ -35,11 +35,10 @@
           </div>
         </div>
 
-
-      </form>
       <button id="userLogin" class="btn waves-effect waves-light" type="submit" name="action">Login
         <i class="material-icons right">control_point</i>
       </button>
+      </form>
     </div>
     <br><br>
