+++ views/newUser.handlebars
@@ -21,7 +21,7 @@
     </div>
     <br><br>
     <div class="row">
-      <form class="col s12">
+      <form id = "newUserForm" class="col s12">
         <div class="row">
           <div class="input-field col s6">
             <input id="first_name" type="text" class="validate">
@@ -50,11 +50,10 @@
             <label for="email">Email</label>
           </div>
         </div>
-
-      </form>
       <button id="newUserSubmit" class="btn waves-effect waves-light" type="submit" name="action">Submit
         <i class="material-icons right">control_point</i>
       </button>
+      </form>
     </div>
     <br><br>
